+++ src/domain/끄적이는/components/History/components/Card/style.css.ts
