+++ src/domain/끄적이는/components/History/components/Column/style.css.ts
