+++ src/domain/끄적이는/components/Today/components/SpellCheckCard/index.tsx
