+++ src/domain/끄적이는/components/Today/components/SpellTypeBox/index.tsx
